+++ jdbc-practices/src/main/java/bookmall/vo/CategoryVo.java
@@ -1,13 +1,13 @@
 package bookmall.vo;
 
 public class CategoryVo {
-	private Long no;
+	private int no;
 	private String name;
 	
-	public Long getNo() {
+	public int getNo() {
 		return no;
 	}
-	public void setNo(Long no) {
+	public void setNo(int no) {
 		this.no = no;
 	}
 	public String getName() {
